grpcapi: clamp the session totals instead of wrapping them
setSessionsTotal wrote `int32(v4 + v6)` on the no-filter path and
`int32(total)` on the filtered-count path, and natSessionCounts
accumulated directly INTO int32 fields. All four wrap negative past
MaxInt32, so the operator-facing "Total sessions" and the NAT
active-translation counts could go negative rather than saturating.

clampInt32 already existed one file over in this same package
(server_nat.go, added by #2282 for exactly this class on the NAT
port-pool size) and simply was not applied here. The session paths now
use it. natSessionCounts widens its accumulator and its per-rule-set map
to int64 and clamps at each protobuf assignment — widening the
accumulator is the load-bearing half, because an int32 field wraps at
increment time, before any clamp at the wire could see it.

Not reachable today: the conntrack tables are far smaller than 2^31 and
the iterators cannot produce more entries than the maps hold. This is
applied because the helper exists, the call sites are one token each, and
an unclamped narrowing conversion is the kind of thing that becomes
reachable when a table bound moves.

Validation: go test -count=1 ./pkg/grpcapi/ passes.
TestSessionsTotalSaturatesInsteadOfWrapping drives the no-filter path
with counts straddling MaxInt32; reverting that call site to a bare
int32() conversion reds it with Total = -2147483648. Narrowing
natSessionCounts.total back to int32 is caught at COMPILE time — the
clampInt32 arguments no longer typecheck — and the test also asserts the
field is wide enough to HOLD an over-range value, so a revert cannot pass
by clamping a counter that already wrapped.

Advances #5250 (A8-b2 F3).

+++ pkg/grpcapi/server_helpers.go
@@ -117,9 +117,14 @@ func (s *Server) telemetry() dataplane.Telemetry {
 
 type natRuleSetKey struct{ from, to string }
 
+// natSessionCounts accumulates in int64 and is CLAMPED at each protobuf
+// assignment (#5250, A8-b2 F3). It used to accumulate directly into int32, so a
+// session table larger than MaxInt32 wrapped the counter negative rather than
+// saturating — the same class #2282 fixed for the NAT port-pool size with
+// clampInt32, which lives one file over and simply was not applied here.
 type natSessionCounts struct {
-	total           int32
-	ruleSetSessions map[natRuleSetKey]int32
+	total           int64
+	ruleSetSessions map[natRuleSetKey]int64
 }
 
 func (s *Server) countSNATSessions(zoneByID map[uint16]string) natSessionCounts {
@@ -132,7 +137,7 @@ func (s *Server) countDNATSessions(zoneByID map[uint16]string) natSessionCounts 
 
 func (s *Server) countNATSessions(flag uint16, zoneByID map[uint16]string) natSessionCounts {
 	counts := natSessionCounts{
-		ruleSetSessions: make(map[natRuleSetKey]int32),
+		ruleSetSessions: make(map[natRuleSetKey]int64),
 	}
 	if !s.dataplaneLoaded() {
 		return counts
